requirements.txt
redis==3.3.11
pika==1.1.0
pyyaml==5.2
flatten-dict==0.2.0

